Issue 85 - Distance and Steps are not sync with the Google Fit
The Google Fitness API has to be enabled in order to work with Google Fitness API. I enabled it on my personal account for this debug build. But you’ll need to setup the fitness api on the VaKidney administrator google account. The steps are available here: https://developers.google.com/fit/android/get-started.

With the changes in this commit and the Google Fit api setup, this issue is fixed.

Please do a fresh install to test this issue.